Tests/LibWeb/Text/expected/ServiceWorker/service-worker-register.txt

ServiceWorker registration failed: SecurityError: Service Worker script response is not a JavaScript MIME type
